A few fixes
