+++ apps/web/package.json
@@ -24,7 +24,7 @@
     "@nextui-org/react": "^2.1.13",
     "@prisma/client": "^5.14.0",
     "@radix-ui/react-avatar": "^1.1.6",
-    "@radix-ui/react-checkbox": "^1.2.2",
+    "@radix-ui/react-checkbox": "^1.2.3",
     "@radix-ui/react-collapsible": "^1.1.8",
     "@radix-ui/react-dialog": "^1.1.11",
     "@radix-ui/react-dropdown-menu": "^2.1.11",
@@ -62,7 +62,6 @@
     "dayjs": "^1.11.10",
     "dompurify": "^3.0.6",
     "eslint": "8.49.0",
-    "eslint-config-next": "13.4.19",
     "firebase": "^11.6.0",
     "firebase-tools": "^13.11.2",
     "flags": "^3.1.1",
